# backendnew/docs/swagger.yaml
basePath: /
definitions:
...
  v1.BusinessUnitResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    required:
    - createdAt
    - id
    - name
    - updatedAt
    type: object
  v1.CreateBusinessUnitRequest:
    properties:
      name:
        example: Finance Department
        type: string
    required:
    - name
    type: object
...
  v1.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    required:
    - code
    - error
    type: object
...
  v1.GenericStatusResponse:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  v1.ListBusinessUnitsResponse:
    properties:
      businessUnits:
        items:
          $ref: '#/definitions/v1.BusinessUnitResponse'
        type: array
      total:
        type: integer
    required:
    - businessUnits
    - total
    type: object
...
  v1.UpdateBusinessUnitRequest:
    properties:
      name:
        example: Updated Finance Department
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A powerful asset discovery and scanning API with Lua script support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Asset Scanner API
  version: "1.0"
paths:
...
  /business-units:
    get:
      consumes:
      - application/json
      description: Retrieve all business units
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListBusinessUnitsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List business units
      tags:
      - business-units
    post:
      consumes:
      - application/json
      description: Create a new business unit with the specified name
      parameters:
      - description: Business unit creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateBusinessUnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.BusinessUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Create business unit
      tags:
      - business-units
  /business-units/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a business unit by its ID
      parameters:
      - description: Business unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GenericStatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete business unit
      tags:
      - business-units
    get:
      consumes:
      - application/json
      description: Retrieve a business unit by its ID
      parameters:
      - description: Business unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.BusinessUnitResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get business unit
      tags:
      - business-units
    put:
      consumes:
      - application/json
      description: Update a business unit's name
      parameters:
      - description: Business unit ID
        in: path
        name: id
        required: true
        type: string
      - description: Business unit update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateBusinessUnitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.BusinessUnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update business unit
      tags:
      - business-units
...
swagger: "2.0"